requirements.txt
# Bot Discord
discord.py>=2.3.0
aiohttp>=3.8.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
lxml>=4.9.0

# Base de données
asyncpg>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilitaires
aiofiles>=23.0.0
python-dateutil>=2.8.0
pytz>=2023.3

# Monitoring et logs
schedule>=1.2.0
colorlog>=6.7.0

# Tests (optionnel)
pytest>=7.4.0
pytest-asyncio>=0.21.0